clean up and separate control server
